+++ src/lib.rs
@@ -10,4 +10,5 @@ pub mod operations;
 pub mod platform;
 mod presentation;
 pub mod resources;
+pub mod session;
 pub mod storage;